feat(det reid): PSTR model
Add of the PSTR model and its base class (Detection ReID).

Also, update of the abstract structure for instance reid detection. We
made the decision to have it implicit so we can keep the original
`InstanceData` as the class. We might use it in the evaluation
function. As it is currently, it only to provide clear information about
the new instance data structure.