feat(docker): working training image
Edit Dockerfile and compose to permit a clean two stages image for training.

The first stage install mmcv (and mmengine) with CUDA support. The image is 14+GB. The second image is taking the compiled version from the first stage and install the mmdet repo on a specific checkout. It "only" takes 8+GB, thanks to the torch image is running on its runtime version.

The size of the images are only illustrative. It heavily depends of the torch base images' sizes.